requirements.txt
python-dotenv~=1.1.0
openai~=1.75.0
requests~=2.32.3
python-telegram-bot~=21.10
Flask~=3.1.0
Werkzeug~=3.1.3
pydantic~=2.11.3